1. divide View into gameView and menuView. 2. fix access to non-existent data when exiting the game.

+++ BadGame/PlayersList.h
@@ -5,7 +5,7 @@
 #include <mutex>
 #include "Clients.h"
 
-const int NUM_OF_DISPLAYED_PLAYERS = 10;
+const size_t NUM_OF_DISPLAYED_PLAYERS = 10;
 
 class PlayersList
 {
